tetragon: Move generic maps in generic_maps.h header
There are several maps used in all generic sensors,
moving them into single generic_maps.h header.
