+++ bpf/process/bpf_generic_kprobe.c
@@ -13,17 +13,11 @@
 #include "types/operations.h"
 #include "types/basic.h"
 #include "generic_calls.h"
+#include "generic_maps.h"
 #include "pfilter.h"
 #include "policy_filter.h"
 
 char _license[] __attribute__((section("license"), used)) = "Dual BSD/GPL";
-
-struct {
-	__uint(type, BPF_MAP_TYPE_PERCPU_ARRAY);
-	__uint(max_entries, 1);
-	__type(key, __u32);
-	__type(value, struct msg_generic_kprobe);
-} process_call_heap SEC(".maps");
 
 int generic_kprobe_setup_event(void *ctx);
 int generic_kprobe_process_event(void *ctx);
@@ -47,44 +41,6 @@ struct {
 		[5] = (void *)&generic_kprobe_output,
 	},
 };
-
-struct {
-	__uint(type, BPF_MAP_TYPE_HASH);
-	__uint(max_entries, 1); // will be resized by agent when needed
-	__type(key, __u64);
-	__type(value, __s32);
-} override_tasks SEC(".maps");
-
-#ifdef __LARGE_BPF_PROG
-struct {
-	__uint(type, BPF_MAP_TYPE_PERCPU_ARRAY);
-	__uint(max_entries, 1);
-	__type(key, __u32);
-	__type(value, struct msg_data);
-} data_heap SEC(".maps");
-#define data_heap_ptr &data_heap
-#else
-#define data_heap_ptr 0
-#endif
-
-struct filter_map_value {
-	unsigned char buf[FILTER_SIZE];
-};
-
-/* Arrays of size 1 will be rewritten to direct loads in verifier */
-struct {
-	__uint(type, BPF_MAP_TYPE_ARRAY);
-	__uint(max_entries, 1);
-	__type(key, int);
-	__type(value, struct filter_map_value);
-} filter_map SEC(".maps");
-
-struct {
-	__uint(type, BPF_MAP_TYPE_ARRAY);
-	__uint(max_entries, 1);
-	__type(key, __u32);
-	__type(value, struct event_config);
-} config_map SEC(".maps");
 
 static struct generic_maps maps = {
 	.heap = (struct bpf_map_def *)&process_call_heap,
